etl/requirements.txt
# Data collection
pytrends>=4.9.0
praw>=7.7.0
google-api-python-client>=2.100.0
requests>=2.31.0

# Data processing
pandas>=2.1.0
numpy>=1.25.0

# Database
supabase>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
tenacity>=8.2.0

# Testing
pytest>=7.4.0
